Adapt `hijack-stream` to Stanford's network
Stanford does some extra window checking that the authors don't
mention. Specifically, it not only checks that the sequence number
is within the correct 2G space, but it also checks that the
acknowlegement number is within 128K of what it should be. This
causes the normal implementation of `hijack-stream` to fail.

This commit adapts `hijack-stream` to send PSH/ACK packets in all
the possible "bins" on Stanford's network.